. F Answer to Day 6 Part 2

+++ src/main/java/org/adventofcode2023/Day6.java
@@ -3,17 +3,17 @@ package org.adventofcode2023;
 import java.util.List;
 
 public class Day6 {
-    public static Integer multipleOfNumberOfWaysToBeatRecord(List<Integer> raceTimes, List<Long> winningDistanceRecords) {
-        int answer = 1;
+    public static Long multipleOfNumberOfWaysToBeatRecord(List<Long> raceTimes, List<Long> winningDistanceRecords) {
+        long answer = 1;
 
         for (int i = 0; i < raceTimes.size(); i++) {
-            int raceTime = raceTimes.get(i);
+            long raceTime = raceTimes.get(i);
             long distanceToBeat = winningDistanceRecords.get(i);
 
             int numberOfWaysToWin = 0;
 
             for (int holdTime = 0; holdTime <= raceTime; holdTime++) {
-                int distanceTraveled = holdTime * (raceTime - holdTime);
+                long distanceTraveled = holdTime * (raceTime - holdTime);
 
                 if (distanceToBeat < distanceTraveled) {
                     numberOfWaysToWin++;
